src/data-storage/data-storage.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { DataStorageService } from './data-storage.service';
import { ConfigModule, ConfigService } from '@nestjs/config';
import { Connection, DataSource, Repository } from 'typeorm';
import * as fs from 'fs';
import * as path from 'path';

// Create a mock entity for testing
class TestEntity {
  id: string;
  name: string;
}

describe('DataStorageService', () => {
  let service: DataStorageService;
  let configService: ConfigService;
  let mockRepository: Partial<Repository<TestEntity>>;
  let mockConnection: Partial<Connection>;
  let mockDataSource: Partial<DataSource>;

  beforeEach(async () => {
    mockRepository = {
      create: jest.fn(),
      save: jest.fn(),
      findOne: jest.fn(),
      find: jest.fn(),
      update: jest.fn(),
      delete: jest.fn(),
    };

    mockConnection = {
      getRepository: jest.fn().mockReturnValue(mockRepository),
      query: jest.fn(),
    };

    mockDataSource = {
      getRepository: jest.fn().mockReturnValue(mockRepository),
      createQueryRunner: jest.fn(),
    };

    const module: TestingModule = await Test.createTestingModule({
      imports: [ConfigModule.forRoot()],
      providers: [
        DataStorageService,
        {
          provide: DataSource,
          useValue: mockDataSource,
        },
        {
          provide: ConfigService,
          useValue: {
            get: jest.fn().mockReturnValue('mock_database_url'),
          },
        },
      ],
    }).compile();

    service = module.get<DataStorageService>(DataStorageService);
    configService = module.get<ConfigService>(ConfigService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });

  describe('createEntity', () => {
    it('should create an entity', async () => {
      const entityData = { name: 'Test Entity' };
      (mockRepository.create as jest.Mock).mockReturnValue(entityData);
      (mockRepository.save as jest.Mock).mockResolvedValue(entityData);

      const result = await service.createEntity(TestEntity, entityData);
      expect(result).toEqual(entityData);
      expect(mockRepository.create).toHaveBeenCalledWith(entityData);
      expect(mockRepository.save).toHaveBeenCalledWith(entityData);
    });
  });

  describe('findEntity', () => {
    it('should find an entity by id', async () => {
      const entityData = { id: '1', name: 'Test Entity' };
      (mockRepository.findOne as jest.Mock).mockResolvedValue(entityData);

      const result = await service.findEntity(TestEntity, '1');
      expect(result).toEqual(entityData);
      expect(mockRepository.findOne).toHaveBeenCalledWith({ where: { id: '1' } });
    });
  });

  describe('backupDatabase', () => {
    it('should create a database backup', async () => {
      const execMock = jest.fn((command, callback) => callback(null, 'Backup successful', ''));
      jest.spyOn(require('child_process'), 'exec').mockImplementation(execMock);

      const backupPath = await service.backupDatabase();
      expect(backupPath).toContain('backups/backup_');
      expect(execMock).toHaveBeenCalled();
    });
  });

  describe('restoreDatabase', () => {
    it('should restore the database from a backup', async () => {
      const execMock = jest.fn((command, callback) => callback(null, 'Restore successful', ''));
      jest.spyOn(require('child_process'), 'exec').mockImplementation(execMock);

      const backupPath = '/path/to/backup.sql';
      await expect(service.restoreDatabase(backupPath)).resolves.not.toThrow();
      expect(execMock).toHaveBeenCalled();
    });
  });
});
